[IMP] web_notify: migration improvements
- Use ES6 in js files
- Update screenshots
- Clean old lint exceptions
- New icon

TT38350
