+++ web_notify/__manifest__.py
@@ -14,8 +14,7 @@
     "depends": ["web", "bus", "base", "mail"],
     "assets": {
         "web.assets_backend": [
-            "web_notify/static/src/js/services/notification.js",
-            "web_notify/static/src/js/services/notification_services.js",
+            "web_notify/static/src/js/services/*.js",
         ]
     },
     "demo": ["views/res_users_demo.xml"],
