+++ web_notify/models/__init__.py
@@ -1,4 +1,1 @@
-# pylint: disable=missing-docstring
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-
 from . import res_users
