+++ web_notify/readme/CONTRIBUTORS.rst
@@ -3,3 +3,6 @@
 * Aitor Bouzas <aitor.bouzas@adaptivecity.com>
 * Shepilov Vladislav <shepilov.v@protonmail.com>
 * Kevin Khao <kevin.khao@akretion.com>
+* `Tecnativa <https://www.tecnativa.com>`_:
+
+  * David Vidal
