+++ web_notify/static/src/js/services/notification.esm.js
@@ -1,0 +1,12 @@
+/** @odoo-module */
+import {Notification} from "@web/core/notifications/notification";
+import {patch} from "web.utils";
+
+patch(Notification.props, "webNotifyProps", {
+    type: {
+        type: String,
+        optional: true,
+        validate: (t) =>
+            ["warning", "danger", "success", "info", "default"].includes(t),
+    },
+});
